+++ turbo_broccoli/custom/__init__.py
@@ -3,12 +3,12 @@
 
 from typing import Any, Callable
 
+from turbo_broccoli.context import Context
 from turbo_broccoli.custom import bytes as _bytes
 from turbo_broccoli.custom import collections as _collections
 from turbo_broccoli.custom import dataclass as _dataclass
 from turbo_broccoli.custom import datetime as _datetime
 from turbo_broccoli.custom import generic as _generic
-from turbo_broccoli.context import Context
 
 try:
     from turbo_broccoli.custom import keras as _keras