+++ turbo_broccoli/custom/collections.py
@@ -4,7 +4,7 @@ from collections import deque, namedtuple
 from typing import Any, Callable, Tuple
 
 from turbo_broccoli.context import Context
-from turbo_broccoli.utils import DeserializationError, TypeNotSupported
+from turbo_broccoli.exceptions import DeserializationError, TypeNotSupported
 
 
 def _deque_to_json(deq: deque, ctx: Context) -> dict: