+++ turbo_broccoli/custom/generic.py
@@ -6,7 +6,7 @@ Serialization of so-called generic object. See
 from typing import Any, Iterable
 
 from turbo_broccoli.context import Context
-from turbo_broccoli.utils import TypeNotSupported
+from turbo_broccoli.exceptions import TypeNotSupported
 
 
 def to_json(obj: Any, ctx: Context) -> dict: