+++ turbo_broccoli/turbo_broccoli.py
@@ -6,7 +6,7 @@ from typing import Any
 
 from turbo_broccoli.context import Context
 from turbo_broccoli.custom import get_decoders, get_encoders
-from turbo_broccoli.utils import TypeIsNodecode, TypeNotSupported
+from turbo_broccoli.exceptions import TypeIsNodecode, TypeNotSupported
 
 
 def _from_jsonable(obj: Any, ctx: Context) -> Any: